+++ src/components/availability-panel/drag-drop-context/index.tsx
@@ -7,7 +7,6 @@ import task from '@/interfaces/task.type';
 import fetch from '@/helperFunctions/fetch';
 import { ASSIGNED } from '@/components/constants/task-status';
 import DroppableComponent from './DroppableComponent';
-import { appendNewTasks } from '@/helperFunctions/getTasks';
 import classNames from '@/components/availability-panel/drag-drop-context/styles.module.scss';
 import { THOUSAND_MILLI_SECONDS, FOURTEEN_DAYS, SECONDS_IN_A_DAY } from '@/components/constants/date'
 
@@ -42,10 +41,23 @@ const DragDropContextWrapper: FC<dragDropProps> = ({
   const [isTaskOnDrag, setIsTaskOnDrag] = useState<boolean>(false);
 
   useEffect(() => {
+    const appendNewTasks = (
+      taskList: task[],
+      unAssignedTasks: task[]
+      ) :task[] => {
+        const oldTasksIds = taskList.map((Task: task) => Task.id);
+        const newTasks = unAssignedTasks.filter((Task: task) => !oldTasksIds.includes(Task.id));
+        return [...taskList, ...newTasks];
+    };
+    const appendNewMembers = (
+      idleMembers: Array<string>,
+      memberList: Array<string>
+      ) :Array<string> => {
+      const newMembers = idleMembers.filter((member) => !memberList.includes(member));
+      return [...memberList, ...newMembers];
+    }
     setTaskList(appendNewTasks(taskList, unAssignedTasks));
-    const newMembers = idleMembers.filter((member) => !memberList.includes(member));
-    const newMemberList = [...memberList, ...newMembers];
-    setMemberList(newMemberList);
+    setMemberList(appendNewMembers(idleMembers, memberList));
   }, [unAssignedTasks, idleMembers]);
 
   const reorder = (list: Array<task | string>, startIndex: number, endIndex: number) => {
